+++ src/pages/signOut/index.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable mobx/missing-observer */
 import { useEffect } from 'react'
 import { useNavigate } from 'react-router-dom'
-import authStore from '@/stores/AuthenticationStore'
+import authStore from '@/stores/AuthStore'
 import { ROUTES } from '@/router/routes'
 
 export const SignOutPage = () => {
