+++ src/router/authorization.tsx
@@ -1,7 +1,7 @@
 import type { FC } from 'react'
 import { Navigate, Outlet } from 'react-router-dom'
 import { observer } from 'mobx-react-lite'
-import authStore from '@/stores/AuthenticationStore'
+import authStore from '@/stores/AuthStore'
 import type { AuthorizationProps } from './types'
 import { ROUTES } from './routes'
